Update local_build.sh
Change the `mkdir` and `chmod` lines to include using the "${RADIUSDESK_VOLUME}" variable, since its used in the rest of the script, and because we're sourcing .env.

Also, there's beyond no need in having `|| exit 1` on every line that does something, especially if we have to run this more than once. Like, if we already have an nginx server running and port 80 is already in use.